pyproject.toml
[project]
name = "web_sanitization"
version = "0.0.0"
description = "Scans LLM outputs for strings to prevent downstream browser script execution"
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8.1"
dependencies = [
    "guardrails-ai>=0.4.0",
    "bleach", 
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "test"
]

[tool.pyright]
include = ["validator"]